未入力データを残すか否かの分岐を入れた。完了ボタン押下時に保存されていなかった問題を解消。残す件数を分岐させるのがまだ。

+++ hakarun/RecordList.swift
@@ -85,7 +85,13 @@ struct RecordList: View {
         // とりあえずハードコーディング
         for index in 5..<vitalrecords.count {
             let delRecord = vitalrecords[index]
-            viewContext.delete(delRecord)
+            if isKeepDone {
+                // 未入力データは残す場合
+                if delRecord.isDone { viewContext.delete(delRecord) }
+            } else {
+                // 未入力データも消す場合
+                viewContext.delete(delRecord)
+            }
         }
         do {
             try viewContext.save()
